VPN-3071(Part 1) Split products (#4839)
* Add new feature inAppProducts
* Split of purchases and products
* This paves the way for purchases on desktop.
